refactor(schedule): move attendance-patch invariant from model to service (#586)
Move ValidateAttendancePatch (a cross-field business invariant: substatus may not be
set while status is expected) out of models/schedule into services/schedule per Rule 12.
The function is pure; all referenced constants/helpers/types stay exported in the model.
Repoint both callers (timetable operations service, api/timetable wrapper) and relocate
the two unit tests to services/schedule.
