+++ backend/models/schedule/instance_student.go
@@ -151,49 +151,6 @@ func (p AttendanceFieldPatch) HasChanges() bool {
 		p.Note != nil || p.NoteClear
 }
 
-// ValidateAttendancePatch enforces attendance patch business rules against the
-// current row so cross-field rules can be evaluated against the final state.
-func ValidateAttendancePatch(patch AttendanceFieldPatch, current *InstanceStudent) []AttendancePatchFieldError {
-	var errs []AttendancePatchFieldError
-
-	if patch.Status != nil && !IsValidAttendanceStatus(*patch.Status) {
-		errs = append(errs, AttendancePatchFieldError{Field: "status", Reason: "must be one of: expected, present, absent"})
-	}
-	if patch.Substatus != nil && !IsValidAttendanceSubstatus(*patch.Substatus) {
-		errs = append(errs, AttendancePatchFieldError{Field: "substatus", Reason: "must be one of: late, excused, sick, field_trip, other"})
-	}
-	if patch.Note != nil && len(*patch.Note) > InstanceStudentNoteMaxLength {
-		errs = append(errs, AttendancePatchFieldError{
-			Field:  "note",
-			Reason: fmt.Sprintf("must be at most %d characters", InstanceStudentNoteMaxLength),
-		})
-	}
-	if len(errs) > 0 {
-		return errs
-	}
-
-	finalStatus := current.Status
-	if patch.Status != nil {
-		finalStatus = *patch.Status
-	}
-
-	finalSubstatusNonNull := current.Substatus != nil
-	if patch.SubstatusClear {
-		finalSubstatusNonNull = false
-	} else if patch.Substatus != nil {
-		finalSubstatusNonNull = true
-	}
-
-	if finalSubstatusNonNull && finalStatus == AttendanceStatusExpected {
-		errs = append(errs, AttendancePatchFieldError{
-			Field:  "substatus",
-			Reason: "cannot be set when status is expected",
-		})
-	}
-
-	return errs
-}
-
 // InstanceStudentRepository defines operations for managing expected/actual
 // attendance on materialized activity instances.
 type InstanceStudentRepository interface {
